requirements.txt
ffmpeg-python
colorlog
requests
torch
PyYAML